Fix issue where some card prices would not return even though a price existed.
This was happening because searching for a card without a set code, will
use any set, and it then searches for the price for that card in that
set, which may not exist. It now falls back to searching for a card
price by only name if it can't get it.